requirements.txt
langchain>=0.3.21
langchain-openai>=0.1.0
pydantic>=2.11.1
python-dotenv>=1.0.1
sqlalchemy>=2.0.30
streamlit>=1.44.0
pymysql>=1.1.0
